% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency}
\alias{efficiency}
\title{Efficiency of an undirected graph}
\usage{
efficiency(g)
}
\arguments{
\item{g}{a graph of type \code{igraph}}
}
\description{
The \code{efficiency}, according to Latora (2001), of an undirected network is calculated by summing the inverse distances
for all node-node pairs of the graph and divide by the same metric for a fully connected network.
}
\references{
Latora, V., & Marchiori, M. (2001). Efficient behavior of small-world networks. Physical review letters, 87(19), 198701.
}
\seealso{
\code{\link{local_efficiency}}
}
