% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_content.R
\name{information_content}
\alias{information_content}
\title{Topological information content}
\usage{
information_content(g)
}
\arguments{
\item{g}{a graph}
}
\description{
Topological information content
}
\details{
The topological information content is defined as
the logarithm of the size of the automorphism group to the base of 2.
}
