% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{local_efficiency}
\alias{local_efficiency}
\title{local efficiency of a graph}
\usage{
local_efficiency(g, order = 1L)
}
\arguments{
\item{g}{a graph}

\item{order}{the order of the egonetwork. The default is taking only direct neighbors.}
}
\description{
\code{local_efficiency} is a generalized form of the measure defined by Latora (2001).
For each node the efficiency of the egonetwork without the central node is calculated. 
The Latora definition uses order = 1 (default).
}
\references{
Latora, V., & Marchiori, M. (2001). Efficient behavior of small-world networks. Physical review letters, 87(19), 198701.
}
\seealso{
Other efficiency measure: \code{\link{efficiency}}
}
\concept{efficiency measure}
