% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{vulnerability_nodes}
\alias{vulnerability_nodes}
\title{Network vulnerability per node}
\usage{
vulnerability_nodes(g, efficiency = NULL)
}
\arguments{
\item{g}{a graph}

\item{efficiency}{A precomputed efficiency for \code{g} (optional)}
}
\description{
The \code{vulnerability}, according to Gol'dshtein (2004) and Latora et al (2005).
}
\references{
Goldshtein, V., Koganov, G. A., & Surdutovich, G. I. (2004). Vulnerability and Hierarchy of Complex Networks. Cond-mat/0409298.

Latora, V., & Marchiori, M. (2005). Vulnerability and protection of infrastructure networks. Physical Review E, 71(1), 015103.
}
