% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{local_efficiency}
\alias{local_efficiency}
\title{local efficiency of a graph}
\usage{
local_efficiency(g, order = 1L)
}
\arguments{
\item{g}{a graph}

\item{order}{the order of the neighborhood. 1 is are direct neighbors.}
}
\description{
For each node te , according to Latora (2001)
}
\seealso{
\code{\link{efficiency}}
}
